fix: correct schema and add test file

+++ db/migrate/20230322001904_create_stars.rb
@@ -3,7 +3,7 @@ class CreateStars < ActiveRecord::Migration[5.2]
     create_table :stars do |t|
       t.string :name
       t.string :type
-      t.integer :age
+      t.float :age
       t.float :size
       t.boolean :can_nova
     end
